+++ r3/core/version
@@ -1,0 +1,1 @@
+2025.0.1
